requirements.txt
asgiref==3.8.1
Django==5.1.2
django-environ==0.12.0
djangorestframework==3.15.2
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.1
psycopg2==2.9.10
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
